Check in a set of CMakeLists.txt files that allowed me to build this package on a test ubuntu machine running buntu 6.8.0-85-generic #85~22.04.1-Ubuntu
This configuration is know to work ONLY IF boost is installed independently
in some directory outside the package build area AND mspass is compiled
and linked against the same set of files.   Linking independently compiled
of even the same version does not seem to work.  To make this work these
environment variables need to be set:
CMAKE_INSTALL_PREFIX
Boost_DIR
Boost_INCLUDE_DIRS

In addition it is more efficient to point at the mspass directory
wher gsl is installed by settting GSL_ROOT_DIR