Cleanup binnedstacker (#7)
* Fix mistakes in control class used in this CLI tool

* Fix errors in load_and_sort function

* Fix bugs in linear_stack and stack_group

* Add new weighting functionality

* Save this mostly working cmake file before making some major changes for consistency with mspass

* Changes to cmake configuration for comptability with mspass changes

* Untested revisions for new pseusosource_stacker CLI tool

* Fix mistake in cmake file add temporary debug lines for top level cmake file

* Modify cmake files to be the same form as mspass

* Check in a set of CMakeLists.txt files that allowed me to build
this package on a test ubuntu machine running buntu 6.8.0-85-generic #85~22.04.1-Ubuntu

This configuration is know to work ONLY IF boost is installed independently
in some directory outside the package build area AND mspass is compiled
and linked against the same set of files.   Linking independently compiled
of even the same version does not seem to work.  To make this work these
environment variables need to be set:
CMAKE_INSTALL_PREFIX
Boost_DIR
Boost_INCLUDE_DIRS

In addition it is more efficient to point at the mspass directory
wher gsl is installed by settting GSL_ROOT_DIR

* Fix a couple bugs in initial debug

* Remove debris from top-level cmake file

* Fix multiple issues found in initial debugging of this new tool

* Change default pf file to match revision to cli tool with the same base name

* Add new tools to CLI install setup and remove confusing pwmig- prefix from names created by previous version

* Fix handling of timing of stacks and add required metadata for output.
This version has been run successfully againt 500,000+ Seismogram objects
so is pretty solid.  To be resolved is if output is reasonable.
Submitting this anyway to assure it builds against most recent mspass
changes

* Change default pf to add required data for new features in last commit

* Fix syntax for github change

* Small change to fix arg parsing mistake

* update the pip command options with -C prefix

---------

Co-authored-by: Ian Wang <22849821+wangyinz@users.noreply.github.com>